chore: build inventory and groups correctly
This commit builds in the inventory and inventory groups correctly even
with the duplicates. In both cases, other columns are concatenated in to
make the non-unique columns (text, code) unique.

It also updates all timestamps to use proper timestamps from the
original database instead of simply defaulting to CURRENT_TIMESTAMP().